requirements.txt
# requirements.txt

dashscope
openai
pyaudio
python-dotenv
numpy
pywebview